+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+### Added
+- SemanticLogger::Log.to_h
+
 ## [4.6.1]
 ### Fixed
 - Avoid leaking sensitive information when logging appender name