+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - SemanticLogger::Log.to_h
+- Fixed #122 make it easier to remove the pid from logs
 
 ## [4.6.1]
 ### Fixed